LayoutTests/platform/blackberry/accessibility/selection-states-expected.txt


This tests that we report the correct selection-related states.


FAIL accessibilityController.focusedElement.isMultiSelectable should be true. Was false.
FAIL accessibilityController.focusedElement.childAtIndex(0).isSelectable should be true. Was false.
FAIL accessibilityController.focusedElement.childAtIndex(0).isSelected should be true. Was false.
FAIL accessibilityController.focusedElement.childAtIndex(1).isSelectable should be true. Was false.
PASS accessibilityController.focusedElement.childAtIndex(1).isSelected is false
PASS accessibilityController.focusedElement.childAtIndex(2).isSelectable is false
PASS accessibilityController.focusedElement.childAtIndex(2).isSelected is false
